+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.9 (2019-10-30)
+Added support to use `ngx-validation-messages` with material ui `mat-form-field` `mat-error` component to display error in standard `ngx-validation-messages` way.
+
+To get more details go to [README.md](https://github.com/lagoshny/ngx-validation-messages#4-with-material-ui-components-using-mat-error-component).
+
 ## 1.0.8 (2019-10-13)
 Updated Angular version.